updating exception tests
Arrived at the removal of Atomic part. Lots of atomic tests failing and probably better to remove now.

[Ticket: #189]
